Add source maps to dll
