+++ lib/webpack/webpack.dll.config.js
@@ -13,6 +13,7 @@ function vendorArray (context) {
 module.exports = function dllConfig (context) {
   const loaders = getLoaders(allLoaders, context);
   return {
+    devtool: 'source-map',
     entry: {
       vendor: vendorArray(context)
     },
@@ -23,8 +24,8 @@ module.exports = function dllConfig (context) {
     },
     plugins: [
       new webpack.DllPlugin({
-	path: path.join(context.projectDir, '.tarec/dll', '[name].manifest.json'),
-	name: '[name]_library'
+        path: path.join(context.projectDir, '.tarec/dll', '[name].manifest.json'),
+        name: '[name]_library'
       })
     ],
     resolve: {
